+++ src/managedcleanroom/azext_managedcleanroom/analytics_frontend_api/aio/operations/_operations.py
@@ -34,11 +34,13 @@ from ...operations._operations import (
     build_collaboration_analytics_queries_document_id_publish_post_request,
     build_collaboration_analytics_queries_document_id_run_post_request,
     build_collaboration_analytics_queries_document_id_runs_get_request,
+    build_collaboration_analytics_queries_document_id_runs_run_id_cancel_post_request,
     build_collaboration_analytics_queries_document_id_vote_post_request,
     build_collaboration_analytics_queries_list_get_request,
     build_collaboration_analytics_runs_job_id_get_request,
     build_collaboration_analytics_secrets_secret_name_put_request,
     build_collaboration_analytics_skr_policy_get_request,
+    build_collaboration_collaborators_get_request,
     build_collaboration_consent_document_id_get_request,
     build_collaboration_consent_document_id_put_request,
     build_collaboration_id_get_request,
@@ -914,6 +916,84 @@ class CollaborationOperations:  # pylint: disable=too-many-public-methods
                 {})  # type: ignore
 
         return cast(JSON, deserialized)  # type: ignore
+
+    @distributed_trace_async
+    async def collaborators_get(self, collaboration_id: str, **kwargs: Any) -> Optional[JSON]:
+        """List all collaborators in a collaboration.
+
+        List all collaborators in a collaboration.
+
+        :param collaboration_id: Required.
+        :type collaboration_id: str
+        :return: JSON object or None
+        :rtype: JSON or None
+        :raises ~azure.core.exceptions.HttpResponseError:
+
+        Example:
+            .. code-block:: python
+
+                # response body for status code(s): 200
+                response == {
+                    "collaborators": [
+                        {
+                            "isOwner": bool,
+                            "objectId": "str",
+                            "tenantId": "str",
+                            "userIdentifier": "str"
+                        }
+                    ]
+                }
+                # response body for status code(s): 422
+                response == {
+                    "loc": [
+                        {}
+                    ],
+                    "msg": "str",
+                    "type": "str"
+                }
+        """
+        error_map: MutableMapping = {
+            401: ClientAuthenticationError,
+            404: ResourceNotFoundError,
+            409: ResourceExistsError,
+            304: ResourceNotModifiedError,
+        }
+        error_map.update(kwargs.pop("error_map", {}) or {})
+
+        _headers = kwargs.pop("headers", {}) or {}
+        _params = kwargs.pop("params", {}) or {}
+
+        cls: ClsType[Optional[JSON]] = kwargs.pop("cls", None)
+
+        _request = build_collaboration_collaborators_get_request(
+            collaboration_id=collaboration_id,
+            api_version=self._config.api_version,
+            headers=_headers,
+            params=_params,
+        )
+        _request.url = self._client.format_url(_request.url)
+
+        _stream = False
+        pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
+            _request, stream=_stream, **kwargs
+        )
+
+        response = pipeline_response.http_response
+
+        if response.status_code not in [200, 404, 422]:
+            map_error(status_code=response.status_code, response=response, error_map=error_map)
+            raise HttpResponseError(response=response)
+
+        deserialized = None
+        if response.content:
+            deserialized = response.json()
+        else:
+            deserialized = None
+
+        if cls:
+            return cls(pipeline_response, deserialized, {})  # type: ignore
+
+        return deserialized  # type: ignore
 
     @distributed_trace_async
     async def invitations_get(
@@ -2998,6 +3078,85 @@ class CollaborationOperations:  # pylint: disable=too-many-public-methods
                 Union[list[JSON], JSON], deserialized), {})  # type: ignore
 
         return cast(Union[list[JSON], JSON], deserialized)  # type: ignore
+
+    @distributed_trace_async
+    async def analytics_queries_document_id_runs_run_id_cancel_post(  # pylint: disable=name-too-long
+        self, collaboration_id: str, document_id: str, run_id: str, **kwargs: Any
+    ) -> JSON:
+        """Cancel query run by run id.
+
+        Cancel query run by run id.
+
+        :param collaboration_id: Required.
+        :type collaboration_id: str
+        :param document_id: Required.
+        :type document_id: str
+        :param run_id: Required.
+        :type run_id: str
+        :return: JSON object
+        :rtype: JSON
+        :raises ~azure.core.exceptions.HttpResponseError:
+
+        Example:
+            .. code-block:: python
+
+                # response body for status code(s): 200
+                response == {
+                    "id": "str",
+                    "status": "str"
+                }
+                # response body for status code(s): 422
+                response == {
+                    "loc": [
+                        {}
+                    ],
+                    "msg": "str",
+                    "type": "str"
+                }
+        """
+        error_map: MutableMapping = {
+            401: ClientAuthenticationError,
+            404: ResourceNotFoundError,
+            409: ResourceExistsError,
+            304: ResourceNotModifiedError,
+        }
+        error_map.update(kwargs.pop("error_map", {}) or {})
+
+        _headers = kwargs.pop("headers", {}) or {}
+        _params = kwargs.pop("params", {}) or {}
+
+        cls: ClsType[JSON] = kwargs.pop("cls", None)
+
+        _request = build_collaboration_analytics_queries_document_id_runs_run_id_cancel_post_request(
+            collaboration_id=collaboration_id,
+            document_id=document_id,
+            run_id=run_id,
+            api_version=self._config.api_version,
+            headers=_headers,
+            params=_params,
+        )
+        _request.url = self._client.format_url(_request.url)
+
+        _stream = False
+        pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
+            _request, stream=_stream, **kwargs
+        )
+
+        response = pipeline_response.http_response
+
+        if response.status_code not in [200, 422]:
+            map_error(status_code=response.status_code, response=response, error_map=error_map)
+            raise HttpResponseError(response=response)
+
+        if response.content:
+            deserialized = response.json()
+        else:
+            deserialized = None
+
+        if cls:
+            return cls(pipeline_response, cast(JSON, deserialized), {})  # type: ignore
+
+        return cast(JSON, deserialized)  # type: ignore
 
     @distributed_trace_async
     async def analytics_datasets_document_id_queries_get(  # pylint: disable=name-too-long